pyproject.toml
[tool.poetry]
name = "eth-transaction-processor"
version = "0.1.0"
description = ""
authors = ["poligrafo"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.112.0"
sqlalchemy = "^2.0.32"
databases = "^0.9.0"
psycopg2-binary = "^2.9.9"
pydantic = "^2.8.2"
web3 = "^6.20.1"
uvicorn = "^0.30.5"
asyncpg = "^0.29.0"
alembic = "^1.13.2"
pytest = "^8.3.2"
httpx = "^0.27.0"
pydantic-settings = "^2.4.0"
hypercorn = "^0.17.3"
aioredis = "^2.0.1"
fastapi-cache2 = "^0.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
